Refactored NameSortCadMap to use TryGetValue to handle when Name info will be treated as a company not a person.
Added SelectAgentNameIdFromMarionAgentImportTable to populate owner.AgentID.

Added ConnectionStringMarion1 if we want to work on the backup of wagapp2_2021_Marion before we practiced on it.

Added messaging classes for when uploads of parts are finished.
Changed [Key] attributes to [ExplicitKey in mCadAccount and mCadOwner.

Corrected all MarionDistributeImport namespaces and usings to MarionUpload.

Renamed Map dictionaries for clarity.